Made it more user-friendly
PyConlangWordGen can now be run without a Python installation or without
passing command line parameters. If it's launched without any
parameters, it will query the user for the pertinent information.

Also added a *.zip file containing the file as a Windows executable,
allowing user to run it without needing to install Python.